requirements.txt
python-telegram-bot==20.8
SQLAlchemy==2.0.28
aiohttp==3.9.3
ari-py==0.3.0
panoramisk==3.10.0
python-dotenv==1.0.1
alembic==1.13.1 